Fix PUT implementation: allow it to create
User was not able to PUT new entry, because the security policy denied
access to unknown ids.
With the new policy we will correctyl return 404 on non existing and
allow to PUT.

This does mean that a user can figgure out if some ID exists in the
system by diferentiating between forbidden and not found.